+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# REPLACE_GLOBALLY_WITH_NEXT_VERSION (3.3.0)
+# 3.3.0
 - PPI-219 - Disable PayPal on Carts with total price of 0
 - PPI-227 - Add possibility to add a suffix to the order number sent to PayPal
 - PPI-281 - Improved storefront behaviour of Express Checkout & Smart Payment Button cancellations and errors
+- PPI-289 - Fix issue where the Express Checkout button could be visible to logged in customers
 - PPI-304 - Fix Smart Payment Buttons being visible with cart errors
 
 # 3.2.1